今天

+++ te/src/main/java/cn/fk/te/product/mapper/ProductMapper.java
@@ -5,5 +5,5 @@ import cn.fk.common.dao.BaseCRUDDao;
 import cn.fk.te.product.model.po.Product;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface ProductMapper extends BaseCRUDDao<Long, Product> {
+public interface ProductMapper extends BaseCRUDDao<Integer, Product> {
 }
